refactor: sqs stable diffusion vision request service 로직 리팩토링
